ss_korean/requirements.txt
absl-py==1.4.0
astor==0.8.1
certifi==2019.11.28
cycler==0.11.0
dataclasses==0.8
future==0.18.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.48.2
h5py==2.10.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib==3.3.4
mne==0.23.4
numpy==1.19.5
opt-einsum==3.3.0
pandas==1.0.1
Pillow==8.4.0
protobuf==3.11.2
pyEDFlib==0.1.15
pyparsing==3.1.1
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.24.2
scipy==1.5.4
six==1.14.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
threadpoolctl==3.1.0
tqdm==4.64.1
typing-extensions==4.1.1
Werkzeug==2.0.3
wrapt==1.16.0
xlrd==1.2.0
zipp==3.6.0
